+++ src/io/javascriptProvider.ts
@@ -1,5 +1,0 @@
-import * as vscode from 'vscode';
-import * as httpyac from 'httpyac';
-export function initJavascriptProvider() {
-  httpyac.io.javascriptProvider.require.vscode = vscode;
-}