+++ src/plugin/index.ts
@@ -1,0 +1,2 @@
+export * from './bailOnFailedTestInterceptor';
+export * from './vscodeHttpyacPLugin';